+++ morpheus-graphql-core/src/Data/Morpheus/Types/Internal/Stitching.hs
@@ -54,6 +54,9 @@ equal err p1 p2
   | p1 == p2 = pure p2
   | otherwise = failure err
 
+fstM :: Applicative m => a -> a -> m a
+fstM x _ = pure x
+
 concatM :: (Applicative m, Semigroup a) => a -> a -> m a
 concatM x = pure . (x <>)
 
@@ -61,17 +64,15 @@ class Stitching a where
   stitch :: (Monad m, Failure ValidationErrors m) => a -> a -> m a
 
 instance Stitching a => Stitching (Maybe a) where
-  stitch Nothing y = pure y
-  stitch (Just x) Nothing = pure (Just x)
-  stitch (Just x) (Just y) = Just <$> stitch x y
+  stitch = optional stitch
 
 instance Stitching (Schema s) where
   stitch s1 s2 =
     Schema
       <$> prop stitch types s1 s2
-      <*> prop stitch query s1 s2
-      <*> prop stitch mutation s1 s2
-      <*> prop stitch subscription s1 s2
+      <*> prop stitchOperation query s1 s2
+      <*> prop (optional stitchOperation) mutation s1 s2
+      <*> prop (optional stitchOperation) subscription s1 s2
       <*> prop stitch directiveDefinitions s1 s2
 
 instance Stitching (TypeLib s) where
@@ -82,6 +83,24 @@ instance Stitching [DirectiveDefinition s] where
 
 instance Stitching [Directive s] where
   stitch = concatM
+
+optional :: Applicative f => (t -> t -> f t) -> Maybe t -> Maybe t -> f (Maybe t)
+optional _ Nothing y = pure y
+optional _ (Just x) Nothing = pure (Just x)
+optional f (Just x) (Just y) = Just <$> f x y
+
+stitchOperation ::
+  (Monad m, Failure ValidationErrors m) =>
+  TypeDefinition c s ->
+  TypeDefinition c s ->
+  m (TypeDefinition c s)
+stitchOperation x y =
+  TypeDefinition
+    <$> prop fstM typeName x y
+    <*> prop fstM typeFingerprint x y
+    <*> prop concatM typeDescription x y
+    <*> prop stitch typeDirectives x y
+    <*> prop stitch typeContent x y
 
 instance Stitching (TypeDefinition cat s) where
   stitch x y =